samples/logging/python/requirements.txt
python-json-logger
prometheus_client
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-flask